Correction in calculateHash()
With the latest version of the MIGS driver, I could not successfully make a payment, due to an Incorrect Hash error that I'm receiving from Omnipay when returning from a completed order.

Upon checking the docs, I found out that there's something missing in the calculateHash() method. The vpc_SecureHash and vpc_SecureHashType parameters should NOT be included in the hash, as indicated in the revised MIGS documentation and sample PHP codes.

On my end, with the proposed changes, I could not complete my order.